improving xcorr for validation

+++ squirrel/library/xcorr.py
@@ -3,8 +3,8 @@ def xcorr(fixed, moving, sigma=1.0):
     from vigra.filters import gaussianGradientMagnitude
     from skimage.registration import phase_cross_correlation
 
-    fixed_ = gaussianGradientMagnitude(fixed, sigma)
-    moving_ = gaussianGradientMagnitude(moving, sigma)
+    fixed_ = gaussianGradientMagnitude(fixed.astype('float32'), sigma)
+    moving_ = gaussianGradientMagnitude(moving.astype('float32'), sigma)
 
     return phase_cross_correlation(
         fixed_, moving_, upsample_factor=10
